requirements.txt
simple-salesforce
pandas
aiohttp
google-auth-oauthlib
google-api-python-client
python-dotenv